fix: pre-render of the `Installation` page
